+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Airbrake Changelog
 
 ### master
 
+### [v8.3.2][v8.3.2] (March 12, 2019)
+
 * Fixed the Rails performance breakdown hook not maintaining performance
   precision ([#936](https://github.com/airbrake/airbrake/pull/936))
 * Fix Rails performance breakdown not being sent if one of the groups is zero
@@ -596,3 +598,4 @@ Airbrake Changelog
 [v8.2.1]: https://github.com/airbrake/airbrake/releases/tag/v8.2.1
 [v8.3.0]: https://github.com/airbrake/airbrake/releases/tag/v8.3.0
 [v8.3.1]: https://github.com/airbrake/airbrake/releases/tag/v8.3.1
+[v8.3.2]: https://github.com/airbrake/airbrake/releases/tag/v8.3.2